requirements.txt
asgiref==3.5.0
astroid==2.11.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
dill==0.3.4
Django==4.0.3
django-phonenumber-field==6.1.0
django-widget-tweaks==1.4.12
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
phonenumbers==8.12.46
Pillow==9.0.1
pip-autoremove==0.10.0
platformdirs==2.5.1
pycodestyle==2.8.0
pylint==2.13.4
pyotp==2.6.0
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
